refactor: apache 패키지가 jwp 패키지를 의존하지 않도록 수정
