+++ tomcat/src/main/java/nextstep/Application.java
@@ -1,5 +1,7 @@
 package nextstep;
 
+import nextstep.jwp.controller.ExceptionControllerFactory;
+import nextstep.jwp.controller.JwpRequestMapping;
 import org.apache.catalina.startup.Tomcat;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -10,7 +12,7 @@ public class Application {
 
     public static void main(String[] args) {
         log.info("web server start.");
-        final var tomcat = new Tomcat();
+        final var tomcat = new Tomcat(JwpRequestMapping.getInstance(), ExceptionControllerFactory.createExceptionControllers());
         tomcat.start();
     }
 }
