+++ tomcat/src/main/java/nextstep/jwp/controller/ExceptionControllerFactory.java
@@ -1,11 +1,9 @@
-package org.apache.catalina;
+package nextstep.jwp.controller;
 
 import java.util.List;
-import nextstep.jwp.controller.InternalServerExceptionController;
-import nextstep.jwp.controller.NotFoundExceptionController;
 import org.apache.coyote.ExceptionController;
 
-public class ControllerFactory {
+public class ExceptionControllerFactory {
 
     public static List<ExceptionController> createExceptionControllers() {
         return List.of(
