+++ tomcat/src/main/java/org/apache/catalina/connector/Connector.java
@@ -7,9 +7,6 @@ import java.net.Socket;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
-import org.apache.catalina.ControllerContainer;
-import org.apache.catalina.ControllerFactory;
-import nextstep.jwp.controller.JwpRequestMapping;
 import org.apache.coyote.Container;
 import org.apache.coyote.http11.Http11Processor;
 import org.slf4j.Logger;
@@ -29,9 +26,8 @@ public class Connector implements Runnable {
     private final ExecutorService executorService;
     private boolean stopped;
 
-    public Connector() {
-        this(new ControllerContainer(new JwpRequestMapping(), ControllerFactory.createExceptionControllers()),
-                DEFAULT_PORT, DEFAULT_ACCEPT_COUNT, DEFAULT_MAX_THREAD_SIZE);
+    public Connector(final Container container) {
+        this(container, DEFAULT_PORT, DEFAULT_ACCEPT_COUNT, DEFAULT_MAX_THREAD_SIZE);
     }
 
     public Connector(final Container container, final int port, final int acceptCount, final int maxThreads) {
