+++ tomcat/src/main/java/org/apache/catalina/startup/Tomcat.java
@@ -1,17 +1,26 @@
 package org.apache.catalina.startup;
 
+import java.io.IOException;
+import java.util.List;
+import org.apache.catalina.ControllerContainer;
+import org.apache.catalina.RequestMapping;
 import org.apache.catalina.connector.Connector;
+import org.apache.coyote.ExceptionController;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.io.IOException;
 
 public class Tomcat {
 
     private static final Logger log = LoggerFactory.getLogger(Tomcat.class);
 
+    private final ControllerContainer container;
+
+    public Tomcat(final RequestMapping requestMapping, final List<ExceptionController> exceptionControllers) {
+        this.container = new ControllerContainer(requestMapping, exceptionControllers);
+    }
+
     public void start() {
-        var connector = new Connector();
+        var connector = new Connector(container);
         connector.start();
 
         try {
